backend/requirements.txt
fastapi
uvicorn
python-jose
pydantic
supabase
python-dotenv
